Create Between Two Sets
Referred to the website: https://programs.programmingoneonone.com/2021/03/hackerrank-between-two-sets-solution.html. Tried to use several iterations of for loops and even a while loop, but the brute force method proved to be more effective for my understanding. Practically got the answer using this website, so no credit goes to me for this.